+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ----------------
 
 - Export plone.app.redirector redirects, if available.
+  Comply with default expectations of redirector section in plone.app.transmogrifier.
   [hvelarde]
 
 - Do not export FormGen and Redirection tools.